feat(admin): Added admin and garage CRUD

+++ server/model/garage.go
@@ -6,7 +6,7 @@ type Garage struct {
 	Latitude      float64 `json:"latitude,omitempty"`
 	Longituted    float64 `json:"longitute,omitempty"`
 	Level         int64   `json:"level,omitempty"`         //level reuired to unlock the garage
-	CoinsRequired int     `json:"coinsRequired,omitempty"` //coins required to unlock the garage
+	CoinsRequired int64   `json:"coinsRequired,omitempty"` //coins required to unlock the garage
 	Locked        bool    `json:"locked,omitempty"`
 }
